requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.6.0
rank-bm25>=0.2.2
PyMuPDF>=1.24.0
python-pptx>=0.6.21
google-adk>=1.0.0
pytest>=8.0.0
